tools/requirements.txt
facebook-business>=19.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
vercel-blob>=0.1.0
